Ensure stability of betamax cassettes
If query parameters are passed as a dict, they can be sent in any order,
which confuses betamax -- it thinks it hasn't seen this request before.
Passing them as a list of tuples means they always get used in the
same order.

Also, sanitise a place the API key was leaking through.