Fix a bug that was causing the 'add book' not to work
Specifically, because I have a bunch of <span>'s in the search results
that didn't have the 'id' attribute, so the event handler would fire.
Logging should make it easier to spot this in future.

Also, make sure the IBSN is actually wrapped in the ISBN class; previously
the <span> was getting closed before I'd added the ISBN value.